Add files via upload
This is a basic version that allows user input to generate a melody and tempo. For now, all notes are considered quarter notes. The major key used is just C major for simplicities sake, and the minor key used is, therefore, A minor. Technically melodies in all the modes can be written regardless of whether major or minor are selected but they are there to relate the input numbers to the scales.